Every function in its own file, part 4 [#CLICKHOUSE-2]
